[size] Remove board config dep from the size checker
The board config is deprecated now. This CL migrates size checker off of
it.

The size checker needs to know both the blobfs layout and the blobfs
compression variables in order to construct blobfs images. Currently,
the compression state is always 'true', so hardcode that for now. The
blobfs layout is passed in as a CLI argument that defaults to "compact".

Change-Id: Id6dee512feb5e7e467f0af2edcc342b18efb4715
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657583
Reviewed-by: Saman Sami <samans@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>