[ffx] Indicate when femu fails to launch
When something goes wrong with launching the emulator process, we want
to inform the user rather than keep them waiting. This CL adds a check
to the "is_running()" function which does a "no-wait" join, to clean up
the process if it's defunct, then checks if it's still there. We then
call the updated function while we're waiting for the emulator to be
"live", and then short-circuit the health check and return an error if
the process is no longer running.

Bug: fxbug.dev/94211
Change-Id: I8b0f8b829d60602830f88d7383e75049a960f22b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659884
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>