[check-licenses] Updates to the project package.
Added the "analyze" file, which loops over all project objects
and searches their licenses & regular files for license texts.

Added the ability to define custom fields in the README.fuchsia files,
which lets us say (for example) all files in the Fuchsia project
must have a copyright header. Or all license files in the Chromium
have a certain format (which helps with analysis).

Reorganized a few of the test artifacts, and other small tweaks
to match the updates in the other packages.

Change-Id: Ic994b31a4b9cd4a4c944e2e2aae4e18cbedaa8dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659383
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>