[netstack3] Ctx type carries separate context
Modify the `Ctx` type to carry separate `EventDispatcher` and
`CoreContext` objects.

This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See fxbug.dev/92605 for more details.

Bug: 92605
Fixed: 93859
Change-Id: I13986d671b905ff1b0b7a5588de8a17ae4bfafa9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640181
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>