[audio][test] Copy hermetic test lib to audio_core/testing/integration
During transition to CFv2 we'll need to keep around an old copy of this
library (CFv1) along with the new copy. Hence we start by making a copy
of the old library. I've moved this to audio_core/testing/integration
since this is a support library for testing audio_core.

This is a pure copy with some minor BUILD.hn refactoring: nothing has
changed except for #include statements.

Bug: 90408
Change-Id: I7474baec40a1cbc4e1e95b27a9662f440dbee822
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659392
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>