[ffx][repo] Disable server by default
This changes ffx repository to initially have the repository server
start in a disabled state. This means that by default when ffx starts
up, it will not try to start the server, even if there are locally
configured repositories in registrations.

Additionally, this also changes `ffx repository server stop` to mark the
server as disabled if we experience any errors when trying to stop the
server. This covers the case where the daemon gets stuck and is unable
to disable the server. With this fix, restarting the daemon should block
the server from starting.

However, if you want the old behavior (where the server is started when
the daemon is started), users should do either:

```
$ ffx repository server start
```

Or the underlying config setting is changed with:

```
$ ffx config set repository.server.enabled true
$ ffx doctor --restart-daemon
```

We'll switch back to not starting the server if the user does:

```
$ ffx repository server stop
```

Or changes the underlying config with:

```
$ ffx config set repository.server.enabled false
$ ffx doctor --restart-daemon
```

Fixed: 95279
Change-Id: I3dfc27f45e20f26ca3294f07ef239c42557c3852
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656487
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>