[size] Do not fail build with budget exceeded
We need to upload the gerrit size report to display the size analysis,
therefore we cannot cause the build step to fail by running the size
checker. When a budget is exceeded, we will print a warning to stdout,
and return exit code 0 (success).

Later, we can give infra an indication of whether the budgets were
exceeded or not.

Bug: 95016
Change-Id: I318fcdc7bd3aa5cde297dd7c4d43997f4168abf8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660023
Reviewed-by: Saman Sami <samans@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>