[netdevice] Remove race possibilities
SessionStarted and SessionStopped paths were releasing and reacquiring
locks, which is breeding ground for races. Remove those with the help of
TA and letting `DeviceInterface` release the locks when it is done
changing the state of the data path.

We still release and reacquire the lock on NotifyDeadSession, which is
less concerning because the lock reacquisition is mostly to deal with
the zombie session, but the data path is already closed.

Multiply: network-device-lib-test: 30
Change-Id: Ie0a35cec24cbdc1a654f2a85d898df37a569fdd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660423
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>