[dhcpv6] Prepare Requesting for Reply to Renew
Currently the `Requesting` state holds `AddressToRequest` address
entries. In an upcoming change processing Reply to Renew, the client
will need to be able to transition to `Requesting` if it receives IAs
with status NoBinding, while also maintaining the existing
AddressEntry::Assigned entries for which the server had extended the
lifetime, as described in RFC 8415 section 18.2.10.1.

In preparation for implementing Reply to Renew, update the `Requesting`
state to hold `AddressEntry` addresses instead of `AddressToRequest`.
This is a mechanical change, `Requesting::addresses` still only contains
`AddressEntry::ToRequest` entries, it's just the type of the address map
that is updated at this stage.

Bug: 76765

Change-Id: I0630a143f9bc55a883109bbb9024756ec5173189
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/654144
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>