[pwauth] Split key enrollment and retrieval traits
This split is necessary to support the different amounts of state that
enrollment vs retrieval flows depend on when using a stateful key
enrollment scheme (like Pinweaver).

This change also pulls apart the implementations of the key sources into
separate files, which helps keep tests closer to the relevant
implementations and allows them to target the appropriate layer of
abstraction.

Change-Id: I9b430b17979605723a777cb3019a2ded16043448
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658791
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>