[kernel][gfx-font] Remove dependency on libgfx-common
There are too many gfx-related libraries, and only one of them needs to
stand apart from the others at the moment: gfx-font-data (renamed in
this CL to gfx-font). Its sole job is to define some default font
configurations for both gigaboot and the kernel - and hence needs to
remain as pure C. Previously it depended on libgfx-common for a single,
simple struct definition, which in turn forced gfx-common to be written
as C (at least in part). So that we can better consolidate and improve
gfx code, this change reverse the dependency between these two
libraries, migrating the simple struct definition over.

Bug: 96043
Change-Id: Iec9859f910472c8487e915980caa66c1095c25a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660814
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>