[trace_processing] Fix check for missing 'fixed' events
The logic in the metric was intended to cleanly return an empty set of
metrics when the data was missing, but would actually throw an exception
in some cases.

In practice, this was never an issue previously as memory_monitor data
would start being collected as the trace started, so there was always a
data point present. With trace slicing, we may slice to a sub-1-second
interval that does not have any samples, so it is important to be able
to cleanly return in these cases.

TEST=fx test sl4f_client_tests
BUG: b/223316328

Change-Id: Id984926bf881f2acac3082947fdb22cb271a3c3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660669
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Brian Hamrick <bhamrick@google.com>
Fuchsia-Auto-Submit: Brian Hamrick <bhamrick@google.com>