[lockdep] No analysis on GuardMultiple::Release
The clang analysis is able to realize that the Guards being released in
GuardMultiple::Release are not held, however the annotation system is
not powerful enough to forward the lock requirements to the function
declaration. As such, we must disable the ananlysis instead.

Change-Id: I481cdbab92bb890c8d84e975020a5c50ec96c3fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661102
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>