[setui] Raise camera device timeout
SetUI waits for a camera to be found for the input API on startup. If we
don't find a camera, we crash, which brings down the system. This is for
privacy reasons to ensure the user has the correct info for the privacy
state.

However, it seems that the 3 second timeout is not always enough as
there aren't any particular guarantees on the camera API side. Raising
the timeout to 30 seconds so that the system shouldn't crash in any
reasonable case.

Bug: 96178
Test: fx test setui_service_tests
Change-Id: I897bc1dea087d6c44e57ae79b9e01252dcf61862
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661195
Reviewed-by: Paul Faria <paulfaria@google.com>
Fuchsia-Auto-Submit: William Xiao <wxyz@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>