[fs_management] add component launching
Add the ability to launch components using fs_management functions
fs_management. If configured with a child_name and optionally a
collection_name, the library will connect to a filesystem component,
possibly launching it if it's not already launched, and use that
component to start the filesystem instance. Mkfs, Fsck, Mount, and
FsInit are all updated with this behavior.

Bug: 91577
Bug: 90698
Test: fx test fs-management-tests
Change-Id: I4657b55f0576bc7c4b05914b8541b2f3f6352c04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641394
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>