[starnix] Add the concept of a "galaxy"
Prior to this change, each component that used Starnix as a runner would
specify how to configure the runtime environment. For example, a
component's .cml would contain a list of filesystem to mount, the kernel
command line, etc. The component would also contain the system image for
the environment in which it was to be run.

This change introduces the concept of a "galaxy." A galaxy is an
instance of the Starnix runner that is packaged with a system image and
a particular configuration (e.g., filesystems to mount, etc.). Each
such instance is exposed as a separate runner. It's important to note
that the starnix runner code and .cml is the same, only the system image
and configuration parameters change between galaxies.

For example, in a future change our gVisor tests will run using the
"stardroid_test" runner, and our chromiumos shell test will run using
the "starmiumos_test" runner (official name TBD).

Change-Id: I4e5d3b7b3063b7ba360f081759c14c658565a1c3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644365
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>