[fxfs] Add support for filtering an iterator
I started looking at the graveyard which had a TODO about using a filter
iterator and then noticed that the major compaction iterator was broken:
it didn't check the first item. It's not a big deal because I don't
think there's a case where we'd ever need to filter the first item, but
this should fix it.

I've chosen to leave the graveyard implementation alone since it's doing
more than just filtering.

Change-Id: Idc6deab1a5e55401e1bfefe892446850ab57cbcb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661282
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Chris Suter <csuter@google.com>