[fidl] Validate empty struct padding byte
The fidlgen_cpp library adds an extra __reserved field to empty structs
which prevents validation. This special cases the coding traits
implementation to support validation of empty struct bytes.

In the future, it is likely desirable to consolidate all empty structs
into a single struct. However, this involves refactoring of the
fidlgen_cpp library and is left to a separate CL.
NaturalEmptyStructCodingTraits will continue to be used once empty
structs are consolidated.

Fixed: 95757
Change-Id: I21f9238d91cf1802eb3efc287cb5aa3a10fb72ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660104
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>