[starnix][proc] Make writing proc directories easier
Abstracts away some of the details of offsets in `readdir` at the cost
of some throwaway computation (generating entries that will be discarded
due to the offset). This is a good trade-off that can be revisited in
the future once performance is deemed to be an issue.

Change-Id: I08970e0a074317c275665589690943a92063a63a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661724
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>