[run-test-suite] Don't check stdio for incomplete test cases
In the case that a test never sends a test case Stopped event,
or no status is contained in the Stopped event, run-test-suite
does not drain stdout and stderr. I argue this is correct - the
other end of the socket may be stuck open, which would cause
run-test-suite to hang if it tries to drain it. Instead, we'll
drop the socket, which can cause missing stdio. This is okay as
we report cases where Stopped is missing as a special error anyway,
so developers should still be able to diagnose the error.

This change updates the tests to not check for exact stdout output to
remove a test flake.

Multiply: run_test_suite_integration_tests
Bug: 93130

Change-Id: I8a5fe67e5572dede8b02112c62a7c00c5f3e6b41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662185
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>