[usb][xhci] Fix request allocation size
In https://paste.googleplex.com/5426217537765376, ASAN detects heap
corruption when an allocation of a `usb_request_t` requests enough
space for a `usb::Request<>`, but needs to account for additional
storage needs of `usb::BorrowedRequest<>`.

These additional storage needs include entry into a list. It's possible
that this results in a loop in the list, which may explain the handle
leaks observed in https://fxbug.dev/96203.

Test: asan no longer detects corruption.
Fixed: 96203
Change-Id: I7c272313050807e48694ce24932057890ae9740c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662267
Reviewed-by: Sam Hansen <hansens@google.com>
Reviewed-by: Ruby Zhuang <rdzhuang@google.com>
Commit-Queue: Devon O'Dell <dhobsd@google.com>