[ffx] Provide hint for finding host logs on internal errors
When ffx encounters an internal error, it cannot parse the error
and report guidance. When the error was caused by a plugin that
uses rust's logging facility, the best place for a developer to
look for more information is ffx host logs. However, the developer
may not know where ffx host logs are stored.

This change reports the host log directories to stderr (with a
fallback to stdout) when ffx encounters an internal error.

Change-Id: I8e46be6fb4b3354b46f7cb984089e07479911b76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659782
Reviewed-by: Andy Weiss <dragonbear@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>