[FuSS] Add fuss_testing in-tree tests to allowlists required.
tqr/511573, as part of its build artifacts, creates a roughly minimal zbi
and blobfs in order to test that ffx scrutiny extract commandline
functionality works as FuSS (which is out-of-tree) expects it to.

This causes us to need to add it to particular allowlists.

b/222339195

Change-Id: I9c1894a8bdcf8eebb8e9b69e4ac7f8bd0fe788c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662231
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Daniel Hwang <arkay@google.com>