[amlogic-video] Add diagnostics and tracing data
1) Added inspect data for amlogic video driver. Now each decoder
will report the allocation of use of the underlying hardware. A
decoder is using the underlying decoder core if IsUtilizingHardware()
returns true.

2) Added more trace data for both the H264 and VP9 decoders and
Amlogic driver. This allows for a more robust picture when executing
traces on the media driver.

3) Added decoder state in trace data. The decoder state will now be
report as a part of trace data. This uses the ASYNC_TRACE_ API which
allows for non-nesting of historical data. This gives use a historical
view of what state the decoder was in at a given time.

4) Added counter to trace decoder usage as a graph in the Perfetto UI.
Currently there is an issue where if two threads are writing data to
the same counter_id it shows up as two distinct values. A bug was logged
and a workaround is implemented in the code. The workaround can be
removed once the issue has been resolved.

Fixed: 69201

Change-Id: I79c3795be520d08b5706c20a0751c6a7a9874477
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/655002
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>