[wlan][policy] replace disconnect list
This change uses PastConnectionData instead of Disconnect data for
penalizing the scores of networks that disconnect shortly after the
connection begins. PastConnectionData contains all the information that
Disconnect data held for saved networks, Disconnect is not needed. The
behavior or network scoring is unchanged by this change.

Bug: 95460
Test: Modified unit tests. Behavior should be the same
Change-Id: I50492e587367fd6e7fea41d596a4eda311e6584b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660429
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>