[hypervisor][arm64] Remove superfluous lock
We previously had to initialise the VPID allocator, and therefore had to
hold a lock to do so. Since that code had been refactored to remove the
Init() function, we no longer need to hold the lock for the allocator.

Bug: 95763
Change-Id: I375960c658082c2d12cd3a426e8520cea1a945dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662383
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>