[audio][test] Move ConfigurePipelineForThermal to parent
For the most part, support for testing audio with the thermal-
mitigation API is defined in HermeticAudioTest, a parent to
HermeticFidelityTest. One method ConfigurePipelineForThermal
is defined in HermeticFidelityTest but is not specific to
the requirements of that test. This CL moves the method from
HermeticFidelityTest to HermeticAudioTest.

Although not at all required for it, this is being done as
part of the work to integrate the new thermal configuration
API.

fx test -do audio_core_google_comms_thermal_tests_astro

Change-Id: I35c50d984d58eee00fe443c14094bab8a3f29f0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662486
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>