[settings] Use BoundedListNode in SettingProxy
In order to reuse a common structure for vectors of inspect-writeable
elements, we switch out usages of VecDeque for inspect with
BoundedListNode, which manages this automatically. An inspect property
"request_number" is introduced here to preserve the previous inspect
keys that tracked the request ordering across setting_types, as the key
is now auto-managed by BoundedListNode.

Bug: 94356
Test: fx test setui_service_tests

Change-Id: I90d10932d26bac1c3f0b0ccdbbbaf2221932c3b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662352
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>