[assembly] Remove the image assembly config input
Now that we have migrated over to product assembly, we no longer need to
upload the image assembly config and it's dependencies, because the
config is generated by the product assembler, and the dependencies are
included in the legacy AIB.

We also include the product assembly config so that it can be update
in-tree and automatically added out-of-tree.

Unfortunately, we cannot simply this to a generated_file(), because we
need to add the custom signing script and the vbmeta keys. In order to
add these files, we have to strip the ../../ prefix, which I do not
believe we can easily do in GN.

Change-Id: I4aebad085f7c0f1491696b250df03a892a67f887
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662804
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>