[lsc] Migrate all clients of fuchsia.sys2.ComponentResolver
This LSC migrates all usages of the `fuchsia.sys2.ComponentResolver`
protocol to `fuchsia.component.resolution.Resolver`. The latter is in
the SDK and the former will be removed after this CL. This was done
in an LSC, instead of a soft transition for the following reasons:

* All usages are in-tree, so we're confident that nothing will break.
* The change is mechanical. That is, it only changes one namespace to
another, and as such it's fair to assume that nothing will break.
* A soft transition would have greatly slowed down the rollout
of `fuchsia.component.resolution.Resolver`.
* It would have been tricky to have `component_manager` support
consuming both protocols at the same time.

Bug: 94581

Change-Id: I0114fae1ef45d35af24571d5df4a28327067b9bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656983
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
API-Review: Shai Barack <shayba@google.com>
Commit-Queue: Yaneury Fermin <yaneury@google.com>