[fshost] Initialize crypt keys on boot
When the system is configured to use Fxfs, initialize the crypt service
with some dummy keys on boot.  This is necessary for using actual crypto
algorithms in Crypt.

The keys themselves are not secure and not intended to be.  A later CL
will switch to reading from a hardware source.

Bug: 94587
Change-Id: Id153c4bce6aa8a07d476b05cdbd03d1abc79e7d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/656493
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>