[rust_crates] Add hmac for usage by password_authenticator
We wish to use HMAC with a known value as a simple one-way function to
produce a low-entropy secret of a known length from the user's password
to use with credential manager in the pinweaver flow.  This helps
minimize the places distance the user's actual password gets transmitted.

Change-Id: Ide8ca440f9947829a04812fa9652d61195df3ec3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662936
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>