[fuchsia.identity.credential] Adjust FIDL as appropriate
`bytes` is already a vector type, so vector<bytes>:1024 becomes a
vector-of-vector-of-byte, which is not what we want.

`reset_secret` is not something we have current plans to expose from
credential_manager at this time; since this is a FIDL table, it's enough
to just note to the reader that the value is not required to be
populated.

Change-Id: I1cd0508e7323c4053e65f06bc758235ad4a3e6e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662933
Reviewed-by: Jody Sankey <jsankey@google.com>
API-Review: Jody Sankey <jsankey@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>