[phys] Refactor phys stdout setup
This separates the code a bit so the uart library is not a direct
dependency of all phys setup.  Instead, it adds generic multiplexing
across serial and graphics.  This is used with the uart library in
existing code, but now in a dependency-injection style.

Bug: 96043
Change-Id: Id0e58e0d04b89161f76c21d09efa69ee138dba25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662904
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>