[audio] Add new thermal schema
The audio subsystem will be moving to a new version of the
API for thermal configuration. As part of this, we will be
changing how we represent our thermal configuration data,
in our static audio_core_config.json file. Essentially, the
"thermal_policy" section will be replaced by a new section
"thermal_states".

This CL adds the "thermal_states" section to the schema, so
that configs (some of which are in other repositories) can
be updated to include both old and new sections. Once this
is complete, AudioCore will change from referencing the old
section, to referencing the new. Once that is done, client
configs can be updated to remove the old "thermal_policy"
section completely.

Bug: 94424

Change-Id: I8ae8e4cb6dd823bd0c07703a20a4ee457d5fba5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662493
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Patrick Shickel <pshickel@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>