[scrutiny] Remove memory leak created by Arc cycle EnginePlugin
The EnginePlugin holds a reference to the PluginManager that holds a
reference to the EnginePlugin when it's registered. This of course is
a cycle which creates a memory leak. Since Scrutiny holds the main
ownership of the Arc<Mutex<PluginManager>>, EnginePlugin now uses
Weak<Mutex<PluginManager>> so EnginePlugin only has a non-owning
reference.

Bug: 88946
Fixed: 96190
Change-Id: Id748e48074ecaeb62ca0f3918c9c14d9301aaaba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661464
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Charles Celerier <chcl@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>