[fdf] Toggle fallback drivers in driver index
The driver index will not return fallback drivers that match if the base
repository has not been resolved. If the `devmgr.require-system` boot
argument is not set or is set to false then driver index will return
fallback drivers that match even if the base repository has not be
resolved.

Bug: 92329
Change-Id: If7a0db90f71bbf30c3fbd24125e7c21d8d62653f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661394
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>