[netemul] Validate length of endpoint name
Because the name of a virtual endpoint in a netemul runner test is
reused for the interface name, in practice all endpoint names must be
<= fuchsia.net.interfaces/INTERFACE_NAME_LENGTH (15) characters long.

Currently, if a test configures an endpoint with a name longer than
that, the netemul client library will truncates the name on endpoint
creation, but the runner will still use the original longer name when
installing an interface in the netstack, causing the test to crash.
Instead, we should eagerly validate the endpoint name and only allow it
if it also works for the interface name.

Test: fx test -o netemul-runner-errors-tests
Change-Id: I24f170a40c93a3a2f7e3e83e34f41cd14a06894e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663207
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>