[fidl] Enable wire format v2 for overnet
In I4f95d937b0d3b828d2a059c7c671157c757f7f4e, support was added for the
FIDL persistence header, which can specify the encoded wire format which
is then used by the decoder to determine how to decode.

This follows on that CL by enabling the FIDL v2 wire format. The v1 wire
format is deprecated.

This should not depend on the CL enabling the persistence header from
rolling everywhere and could have been part of that CL. It is separated
out out of an abundance of caution.

Change-Id: Ie8e764afff87c1f9a38af8b4771b32fc2d2bcc4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663204
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>