[pwrbtn-monitor] Prefactor for OOM monitor changes
Refactor code so that more happens on the async loop. This moves
the device watching code to an async task. Previously this code
ran before the async loop started. This code waits forever on
targets without a power button.

This prefactor prepares for expanding the job of pwrbtn-monitor
to also listen for kernel OOM events. In a future change we'll
run the button and OOM listening code concurrently on the same
loop.

Bug: 91708
Change-Id: I2c2a6d2b3568feda3880f91573ca0b7e9f51d68e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660807
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Justin Mattson <jmatt@google.com>