[wlan] Use security authenticators in SME clients.
This change uses the `Authentication` FIDL message with bespoke
negotiation in code that communicates directly with SME (rather than the
Policy layer). `SecurityContext` types have been introduces in the
testing service utilities and the `wlandev` tool, which are used to
negotiate authentication using bespoke heuristics that maintain current
behavior in SME. Note that this is a preliminary change and should not
alter the behavior of the testing infrastructure nor `wlandev`.

This change was factored out of a larger abandoned change: fxr/643789.

Testing: (Refactored) unit tests and manual testing.
Change-Id: I481246524ac86154c7b0a38f79d082cddfbfecfc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660009
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>