[kernel][vm] EnumerateChildren return zx_status_t
Changes EnumerateChildren, and its locked/internal variants, to return
an explicit zx_status_t instead of a bool. The intention is to allow
for disambiguating failures, specifically the case where the aspace is
in the destroyed state.

The actual VmEnumerator interface is unchanged and still returns bool
as there are not yet meaningful errors to expose here to justify
updating all the implementors.

Bug: 94078
Change-Id: I66b3df1cf5165428865d0b1de182fca3047ef22c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660523
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>