Ensure public_deps is used to depend on rustc_library when appropriate.
Ensure our build system will work with upstream GN.

A few changes that will allow us to use upstream GN in the Fuchsia build
(we are currently stuck with an older revision due to a breaking
upstream GN change [1], see bug for details).

This ensures that `public_deps` is used instead of `deps` in group() and
source_set() targets that depend on rust_library() instances. Without
this their dependent cannot import the corresponding crate anymore.

Note that for all of this to work properly, the CL at [2] and its parent
must be submitted to GN first. I have verified manually that this CL fixes
the Fuchsia build in this case.

Bug: 95999

[1] https://gn-review.googlesource.com/c/gn/+/13060
[2] https://gn-review.googlesource.com/c/gn/+/13280

Change-Id: I521fab7c4d273cdc980b4caa5fb0104587ef8cc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660607
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>