[kernel][job] Unlocked JobEnumerator
Changes JobDispatcher::EnumerateChildren to, if not recursing, call the
JobEnumerator methods without the parent JobDispatcher::lock_ held. Not
holding the lock has the effect of allowing the ZX_INFO_JOB_* queries
to be able to user copy without holding a lock. The actual set of
children enumerated remain a consistent snapshot as the set is gathered
with the lock held.

With the change to EnumerateChildren the method is split into two
methods, one for recursing and for direct descendants, instead of using
a bool since both the implementation and locking behavior are
sufficiently different for recursion.

Bug: 94078
Change-Id: Ia2f41f48582af718ca4d7ae08ab667106fc73073
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660525
Reviewed-by: Carlos Pizano <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>