[rndis-function] Drop lock before unbind callback
Once the unbind callback is triggered, the driver framework may call
release which deletes the driver. If we still hold the lock while this
happens then there is a use after free when release is called. We don't
need to hold the lock anyway, the callback doesn't need to be protected
and we can safely remove the debug assert (which did need the lock.)

Fixed: 95777

Change-Id: Icd6c4c9158fa9d5b218443493cbcf629197950cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662632
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>