[swd-e2e-tests] Fix workstation e2e reboot tests
Previously, we tried to add a dependency on the target-side sl4f, which
the ota and reboot tests needs to work. Unfortunately the way we added
this only depended on the host-side parts of the sl4f package, which is
empty. This worked previously since we indirectly picked up a dependency
on sl4f from another target, which has since been removed.

This patch then adds an explicit dependency on sl4f that's not
parameterized by what platform we're building on, which fixes the
dependency. This gets the reboot test to pass, but the OTA tests are
blocked by other issues.

Bug: 96481
Change-Id: I94eba273967d258bd5730819fcd26eb5a9dc1224
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663423
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>