[carnelian] Add a way to queue a message for a facet
In a recent discussion of the implementation of a progress bar,
it became clear that the plumbing needed to send a message
from a closure to a facet made it hard to build facets that can
do some work independently of the hosting view controller.

This change is a small that that makes matters a bit better. So long
as the facet is given a cloned app sender and records its facet ID, it
can send itself messages.

Testing: manual with the spinning square example.

Change-Id: I460f3c9c64bc256cccc9dfbad1b2111dd2c281ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663118
Reviewed-by: Dragoș Tiselice <dtiselice@google.com>
Fuchsia-Auto-Submit: Rob Tsuk <robtsuk@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>