[vim3][ethernet] Move WOL enable from Aml eth driver to vim3 MCU driver
Currently both the Amlogic ethernet board driver and the vim3 MCU driver
bind to a single I2C. This I2C is needed for one register write during
ethernet initialization, and is the only I2C that requires multi bind.
Move this call to the MCU driver so that multi bind can be disabled in
the I2C core driver.

Bug: 96293
Test: Connected to vim3 via ethernet, used dm dump to verify that only
      the MCU driver bound to this I2C
Test: fx test vim3-mcu-test
Test: Ran device-enumeration-test on vim3
Change-Id: I9fdb1098ceb7289f2c5988fcaa06ad033351b281
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663402
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>