[ffx] Refactor target wait.
This refactors target wait so that multiple points of failure can be
captured and store in logs, as well as provide a simple way to verify
that the target is discovered, and RCS is available.

The motivation for this change is in integration testing.

Bug: 95702

Change-Id: Ic46a21ff087c5008bbefccee1695e26e17c0e5ae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/663643
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Amit Uttamchandani <amituttam@google.com>