[kernel][vm] Do not collect loaned pages on write
LookupPagesLocked should not collect loaned pages if
PrepareForWriteLocked was called successfully. If PrepareForWriteLocked
encounters any loaned pages, it will terminate the page walk, so to
ensure that exactly the same number of pages were dirtied as being
returned by LookupPagesLocked, collect_pages needs to terminate if it
encounters a loaned page.

This scenario would previously be caught by a debug assert that ensures
that the number of pages dirtied matches the number of pages looked up.

Bug: 63989
Change-Id: Iadd8e8c5392cd803755aa0c8219ae6c8847dd49a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662708
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>