[network/tests] Rewrite forwarding test to use sandbox
The test currently uses the netemul runner. It doesn't use any features
of the runner that the sandbox doesn't fulfill, and with the sandbox it
requires less configuration, is easier to parameterize over endpoint
type and netstack version, and is more in line with the rest of the
netstack integration tests.

Bug: 96132
Fixed: 96458
Fixed: 96459
Test: fx test netstack-forwarding-integration-test
Change-Id: I79a335fd8d6852a72b6d268a3b9cde828ad95ca6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/662827
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>