[fidl][gidl] Validate malformed inline markers
Bit 0 of the envelope flag field is used to determine whether or not the
envelope is inlined or not. The other bits are unused and previously
unchecked in most bindings. However, a fuzzer run caught the lack of
validation of the unused bits, but all FIDL bits should be validated.

Fixed: 96094
Change-Id: I9b4d33ab47378ed24d8de3d485c6392d92da8e15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/661042
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
API-Review: Pascal Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>